docking request changed exhaustiveness type to int

+++ lib/features/docking/data/models/docking_request_model.dart
@@ -12,7 +12,7 @@ class DockingRequestModel {
   final double sizeX;
   final double sizeY;
   final double sizeZ;
-  final double exhaustiveness;
+  final int exhaustiveness;
 
   const DockingRequestModel({
     this.proteinFilePath,
